fix: codium binary now pid 1
The container engine relies on the application being pid 1 to know when to terminate the container. Using Microsoft's wrapper script made it difficult to keep the container alive as codium was not pid 1. So now the binary is executed directly by the container engine after a handoff from the entrypoint.sh